+++ test/async/abort_on_drop.cpp
@@ -35,7 +35,7 @@ auto ready_value() -> async::coro<int> {
 auto wait_for(std::atomic<bool>& value) -> bool {
     for (int i = 0; i < 1000; ++i) {
         if (value.load(std::memory_order_acquire)) return true;
-        thread::sleep(time::Duration::from_millis(1));
+        thread::sleep(time::Duration::from_millis(u64(1)));
     }
     return false;
 }
@@ -50,7 +50,7 @@ TEST(RstdAsyncAbortOnDrop, AwaitsCompletedTask) {
 }
 
 TEST(RstdAsyncAbortOnDrop, DropAbortsPendingTaskAfterMoves) {
-    auto runtime = async::RuntimeBuilder::multi_thread().worker_threads(1).build().unwrap();
+    auto runtime = async::RuntimeBuilder::multi_thread().worker_threads(usize(1)).build().unwrap();
     auto started = std::atomic<bool> { false };
     auto dropped = std::atomic<bool> { false };
 
@@ -66,7 +66,7 @@ TEST(RstdAsyncAbortOnDrop, DropAbortsPendingTaskAfterMoves) {
 }
 
 TEST(RstdAsyncAbortOnDrop, MoveAssignmentAbortsPreviousTask) {
-    auto runtime       = async::RuntimeBuilder::multi_thread().worker_threads(1).build().unwrap();
+    auto runtime = async::RuntimeBuilder::multi_thread().worker_threads(usize(1)).build().unwrap();
     auto first_started = std::atomic<bool> { false };
     auto first_dropped = std::atomic<bool> { false };
     auto next_started  = std::atomic<bool> { false };
@@ -90,7 +90,7 @@ TEST(RstdAsyncAbortOnDrop, MoveAssignmentAbortsPreviousTask) {
 }
 
 TEST(RstdAsyncAbortOnDrop, DetachKeepsTaskRunning) {
-    auto runtime = async::RuntimeBuilder::multi_thread().worker_threads(1).build().unwrap();
+    auto runtime = async::RuntimeBuilder::multi_thread().worker_threads(usize(1)).build().unwrap();
     auto started = std::atomic<bool> { false };
     auto dropped = std::atomic<bool> { false };
     auto guarded = async::AbortOnDropHandle { runtime.spawn(pending_task(started, dropped)) };
